:add - Diogo Main Code
Added my main file, which i could not export to PDF sadly (so Joel needs to do it), which contains a hopefully more abbreviated version of the previous code while also maintaining its fidelity. I did remove a few details however, so previous code is still more "in depth" (since mine removes the graphs it also does not help this).